Hardcode pinned tweet ID to not delete

+++ main.go
@@ -50,11 +50,13 @@ func deleteFromTimeline(api *anaconda.TwitterApi, ageLimit time.Duration) {
 			log.Error("Couldn't parse time ", err)
 		} else {
 			if time.Since(createdTime) > ageLimit {
-				_, err := api.DeleteTweet(t.Id, true)
-				log.Info("DELETED: Age - ", time.Since(createdTime).Round(1*time.Minute), " - ", t.Text)
-				if err != nil {
-					log.Error("Failed to delete! ", err)
-				}
+                if t.Id != "975763655773089797" {
+                    _, err := api.DeleteTweet(t.Id, true)
+                    log.Info("DELETED: Age - ", time.Since(createdTime).Round(1*time.Minute), " - ", t.Text)
+                    if err != nil {
+                        log.Error("Failed to delete! ", err)
+                    }
+                }
 			}
 		}
 	}
